pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.4.0"
description = "Page Loader (Hexlet Python Project Level 3)"
authors = ["neihaoo <neihaoo@gmail.com>"]
repository = "https://github.com/neihaoo/python-project-lvl3"
classifiers = [
  "Topic :: Utilities"
]
packages = [
  { include = "page_loader" },
]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.10.0"
progress = "^1.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
wemake-python-styleguide = "^0.16.1"
requests-mock = "^1.9.3"

[tool.poetry.scripts]
page-loader = "page_loader.scripts.page_loader:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
